@@ -4,20 +4,19 @@
   "type": "symfony-bundle",
   "license": "MIT",
   "require": {
-    "php": ">=7.4|^8.0",
+    "php": "^8.2",
     "ext-soap": "*",
-    "marc-mabe/php-enum": "^3.0|^4.0",
-    "symfony/http-kernel": "^5.4|^6.0",
-    "webmozart/assert": "^1.3"
+    "symfony/http-kernel": "^6.0|^7.0",
+    "webmozart/assert": "^1.11"
   },
   "require-dev": {
     "roave/security-advisories": "dev-master",
-    "phpunit/phpunit": "^9.5",
-    "symfony/phpunit-bridge": "6.1.*",
-    "phpro/grumphp": "^1.5.0",
-    "friendsofphp/php-cs-fixer": "^3.4",
-    "phpstan/phpstan": "^1.4",
-    "phpstan/phpstan-webmozart-assert": "^1.0"
+    "phpunit/phpunit": "^10.5",
+    "symfony/phpunit-bridge": "6.1.*|^7.0",
+    "phpro/grumphp": "^2.8",
+    "friendsofphp/php-cs-fixer": "^3.64",
+    "phpstan/phpstan": "^1.12",
+    "phpstan/phpstan-webmozart-assert": "^1.2"
   },
   "autoload": {
     "psr-4": {
