@@ -4,15 +4,9 @@ declare(strict_types=1);
 
 namespace Answear\MeestBundle\Response;
 
-class GetApiVersion implements ResponseInterface
+readonly class GetApiVersion implements ResponseInterface
 {
-    /**
-     * @var string
-     */
-    public $return;
-
-    public function __construct(string $return)
+    public function __construct(public string $return)
     {
-        $this->return = $return;
     }
 }
