@@ -8,8 +8,5 @@ use Answear\MeestBundle\Response\DTO\CityDTO;
 
 class SearchCityByPostCode implements ResponseInterface
 {
-    /**
-     * @var CityDTO
-     */
-    public $return;
+    public CityDTO $return;
 }
