@@ -4,31 +4,11 @@ declare(strict_types=1);
 
 namespace Answear\MeestBundle\Service;
 
-class ConfigProvider
+readonly class ConfigProvider
 {
-    /**
-     * @var string
-     */
-    private $apiUrl;
-
-    /**
-     * @var string
-     */
-    private $apiKey;
-
-    public function __construct(string $apiUrl, string $apiKey)
-    {
-        $this->apiUrl = $apiUrl;
-        $this->apiKey = $apiKey;
-    }
-
-    public function getApiUrl(): string
-    {
-        return $this->apiUrl;
-    }
-
-    public function getApiKey(): string
-    {
-        return $this->apiKey;
+    public function __construct(
+        public string $apiUrl,
+        public string $apiKey,
+    ) {
     }
 }
